Optimize computation of `in_playlist`
Every time `in_playlist` is written on a track, we recompute the
corresponding album field thanks to `_inverse_in_playlist`.

However, when adding several tracks to a playlist, each playlist line
creation triggers a `write` on the track. This means that if there are X
tracks on an album, we execute `_inverse_in_playlist` X times.

To avoid this, we move the call of `_inverse_in_playlist` at the end of
the tracks addition, so we are sure it is recomputed only once. We also
add when unlinking playlist lines or changing the current playlist.